add page-side component
