+++ src/components/book/page.tsx
@@ -1,5 +1,6 @@
 import React from "react"
-import {front, back, paper,  pageCornerForward, pageCornerBackward, frontContent, backContent,  flipped, tag, tagBack} from "../../styles/book.module.css"
+import {paper, flipped,} from "../../styles/book.module.css"
+import PageSide from "./page-side"
 import { PageContent } from "../../types"
 
 type PageProps = {
@@ -11,27 +12,12 @@ type PageProps = {
     goToPage: Function 
 }
 
-
-
-const tagColours = ["red", "orange", "blue"]
-
-
 const Page = ({data, currentLocation, zIndex, index, tagIndex, goToPage}: PageProps) => {
   return (
         <div className={`${paper} ${currentLocation >= index + 3 ? flipped : null}`}
     style={{zIndex: zIndex}}>
-        <div className={`${front} ${currentLocation === index + 2 ? pageCornerForward : null } ${currentLocation === index + 3 ? pageCornerBackward : null }`}>
-            {data.has_tag ?<div className={tag}  onMouseDown={() => goToPage(index + 2)} style={{backgroundColor: tagColours[tagIndex], marginLeft: `${(tagIndex + 1) * 6}%`}}></div> : null}
-            <div className={frontContent}>
-                <h1>{data.front}</h1>
-            </div>
-        </div>
-        <div className={`${back} ${currentLocation === index + 2 ? pageCornerForward : null } ${currentLocation === index + 3 ? pageCornerBackward : null }`}>
-            {data.has_tag ?<div className={tagBack}  onMouseDown={() => goToPage(index + 2)} style={{backgroundColor: tagColours[tagIndex], marginLeft: `${(tagIndex + 1) * 6}%`}}></div> : null}
-            <div className={backContent}>
-                <h1>{data.back}</h1>
-            </div>
-        </div>
+        <PageSide tagIndex={tagIndex} index={index} goToPage={goToPage} currentLocation={currentLocation} data={data}/>
+        <PageSide tagIndex={tagIndex} index={index} goToPage={goToPage} currentLocation={currentLocation} data={data} isBack={true}/>
     </div>
   )
 }
